requirements.txt
fastapi[standard]
python-dotenv
sqlmodel